Move system-dependant loading from geos.py to libgeos.py, check GEOS versions from both geos-config and the dynamic library.
